@@ -1,17 +1,18 @@
-name: openfoodfacts
+name: api_formation
 description: A web server built using the shelf package.
 # version: 1.0.0
 # homepage: https://www.example.com
 # author: g123k <email@example.com>
 
 environment:
-  sdk: '>=2.6.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
-  args: ^1.4.2
-  shelf: ^0.7.2
-  http: ^0.12.2
-  queries: ^0.1.9
+  args: ^2.3.0
+  path: ^1.8.1
+  shelf: ^1.3.0
+  http: ^0.13.4
+  flinq: ^2.0.2
 
 dev_dependencies:
   pedantic: ^1.11.0
