@@ -1,20 +1,20 @@
+import '../../../../../utils/text_utils.dart';
 import '../../../../data/ingredients_list.dart';
-import '../../../../utils/text_utils.dart';
 
-class ProductTraces {
-  final List<String> list;
-  final List<_TraceItem> details;
+class ProductAllergens {
+  final List<String>? list;
+  final List<_TraceItem>? details;
 
-  ProductTraces.fromAPI(Map<String, dynamic> api, String lng)
-      : list = _extractList(api['traces_tags'], lng),
-        details = _extractDetails(api['traces_tags']);
+  ProductAllergens.fromAPI(Map<String, dynamic> api, String lng)
+      : list = _extractList(api['allergens_tags'], lng),
+        details = _extractDetails(api['allergens_tags']);
 
-  static List<String> _extractList(Object tags, String lng) {
+  static List<String>? _extractList(dynamic tags, String lng) {
     if (tags == null || tags is! List) {
       return null;
     }
 
-    List<String> list = [];
+    var list = <String>[];
 
     for (var tag in tags) {
       var translations = ingredientsTranslations(tag);
@@ -30,12 +30,12 @@ class ProductTraces {
     return list;
   }
 
-  static List<_TraceItem> _extractDetails(Object tags) {
+  static List<_TraceItem>? _extractDetails(dynamic tags) {
     if (tags == null || tags is! List) {
       return null;
     }
 
-    List<_TraceItem> list = [];
+    var list = <_TraceItem>[];
 
     for (var tag in tags) {
       list.add(_TraceItem.fromAPI(tag));
@@ -44,17 +44,17 @@ class ProductTraces {
     return list;
   }
 
-  Map<String, Object> toJson(String language) {
-    if (list == null || list.isEmpty) {
+  Map<String, Object>? toJson(String language) {
+    if (list == null || list!.isEmpty) {
       return null;
     } else if (TextUtils.isNotEmpty(language)) {
       return {
-        'list': list,
+        'list': list!,
       };
     } else {
       return {
-        'list': list,
-        'details': details?.map((i) => i.toJson())?.toList(growable: false)
+        'list': list!,
+        'details': details!.map((i) => i.toJson()).toList(growable: false)
       };
     }
   }
@@ -68,6 +68,6 @@ class _TraceItem {
 
   Map<String, Object> toJson() => {
         'translations':
-            translations?.map((t) => t.toJson())?.toList(growable: false)
+            translations.map((t) => t.toJson()).toList(growable: false)
       };
 }
