@@ -5,7 +5,12 @@ import 'package:args/args.dart';
 import 'server.dart';
 
 void main(List<String> args) async {
-  var parser = ArgParser()..addOption('port', abbr: 'p', defaultsTo: '8080');
+  var parser = ArgParser()
+    ..addOption(
+      'port',
+      abbr: 'p',
+      defaultsTo: '8080',
+    );
 
   var result = parser.parse(args);
 
@@ -13,7 +18,8 @@ void main(List<String> args) async {
 
   if (port == null) {
     stdout.writeln(
-        'Could not parse port value \'${result['port']}\' into a number.');
+      'Could not parse port value \'${result['port']}\' into a number.',
+    );
     // 64: command line usage error
     exitCode = 64;
     return;
